requirements.txt
streamlit
ultralytics
opencv-python
opencv-contrib-python
deepface
tensorflow
torch
torchvision
pandas
numpy
plotly
openai
librosa
soundfile
pyyaml
python-dotenv
tqdm
Pillow
fer
scikit-learn
scipy
matplotlib
seaborn
python-ffmpeg
colorama
psutil